+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project changes will be shown here.
 ## [unreleased]
 ### Fixed
 - Issue 315 - Monitored SR formatting issues.
+### Added
+- Issue 317 - Customer 3rd party contact information
 
 ## [v1.10.1] - 2018-08-03
 ### Fixed